Added bch support and moved to Double instead of Float

+++ src/Binding/Ark.hs
@@ -9,7 +9,7 @@ import Data.Text
 
 data Ark = Ark
   { name :: Text,
-    balance :: IO (Maybe Float),
+    balance :: IO (Maybe Double),
     address :: Text
   }
 
@@ -24,7 +24,7 @@ getArkHolding addr
           balance = getArkBalance,
           address = addr }
     where
-      getArkBalance :: IO (Maybe Float)
+      getArkBalance :: IO (Maybe Double)
       getArkBalance = do
         r <- get ("https://explorer.ark.io:8443/api/accounts?address=" ++ (unpack addr))
         return $ ((/10e7) . read . unpack) <$> (r ^? responseBody . key "account" . key "balance" . _String)
