+++ src/Binding/Btc.hs
@@ -9,7 +9,7 @@ import Network.Wreq (responseBody, get)
 
 data Btc = Btc
   { name :: Text,
-    balance :: IO (Maybe Float),
+    balance :: IO (Maybe Double),
     address :: Text
   }
 
@@ -23,7 +23,7 @@ getBtcHolding addr
   = Btc { name = "Btc",
           balance = getBtcBalance,
           address = addr }
-  where getBtcBalance :: IO (Maybe Float)
+  where getBtcBalance :: IO (Maybe Double)
         getBtcBalance = do
           r <- get $ unpack $ intercalate "" ["https://blockchain.info/address/", addr, "?format=json"]
           return $ ((/1e8) . fromIntegral) <$> (r ^? responseBody . key "final_balance" . _Integer)
