+++ src/Binding/Xrp.hs
@@ -9,7 +9,7 @@ import Network.Wreq (responseBody, get)
 
 data Xrp = Xrp
   { name :: Text,
-    balance :: IO (Maybe Float),
+    balance :: IO (Maybe Double),
     address :: Text
   }
 
@@ -23,7 +23,7 @@ getXrpHolding addr
   = Xrp { name = "Xrp",
           balance = getXrpBalance,
           address = addr }
-  where getXrpBalance :: IO (Maybe Float)
+  where getXrpBalance :: IO (Maybe Double)
         getXrpBalance = do
           r <- get $ unpack $ intercalate "" ["https://data.ripple.com/v2/accounts/", addr, "/balances/?currency=XRP"]
           return $ (read . unpack) <$> (r ^? responseBody . key "balances" . nth 0 . key "value" .  _String)
